add regex check for link and print only slow actions

+++ assignment2/version2/init.go
@@ -29,7 +29,7 @@ type outputConfig struct {
 	Seedfile          string
 	ParsingResultFile string
 
-	SlowAction float64
+	SlowAction int64
 }
 
 type systemConfig struct {
